codesplit components/dashboard, merge to prod
