+++ components/dashboard/context/ApiProvider.tsx
@@ -1,6 +1,7 @@
 import { Directory, Folder } from 'lib/models';
-import { DashboardApiContext, useDashboard } from './Context';
+import { DashboardApiContext } from './Context';
 import React from 'react';
+import { useDashboard } from './hooks';
 
 /**
  * Wrapper component defining DashboardContext and managing UI state updates
@@ -14,11 +15,25 @@ import React from 'react';
  */
 
 export const DashboardApiProvider = ({ children, ...rest }) => {
-  const { kernel } = useDashboard();
+  const { kernel, selected } = useDashboard();
 
   const [isPending, startTransition] = React.useTransition();
 
-  const [clipboard, setClipboard] = React.useState<Folder[]>([]);
+  const [clipboard, setClipboardState] = React.useState<Folder[]>([]);
+
+  const [expanded, setExpandedState] = React.useState(false);
+
+  const [recentAction, setRecentAction] = React.useState<
+    'rename' | 'move' | 'copy' | 'delete' | null
+  >(null);
+
+  const [destination, setDestination] = React.useState<Directory>(null);
+
+  const setClipboard = (folders: Folder[]) =>
+    startTransition(() => setClipboardState(folders));
+
+  const setExpanded = (state: boolean) =>
+    startTransition(() => setExpandedState(state));
 
   // Listen for select event
 
@@ -77,7 +92,7 @@ export const DashboardApiProvider = ({ children, ...rest }) => {
           kernel.trigger('error', 'Missing source or destination Folders');
         }
         kernel.trigger('cut', []);
-        kernel.trigger('select', null)
+        kernel.trigger('select', null);
       }
     );
 
@@ -106,10 +121,88 @@ export const DashboardApiProvider = ({ children, ...rest }) => {
     return () => createEvent.unsubscribe();
   }, [kernel]);
 
+  const handleActionExpand = (e: React.SyntheticEvent) => {
+    if (expanded) setRecentAction(null);
+    setExpanded(expanded ? false : true);
+  };
+
+  const handleCut = (e: React.SyntheticEvent) => {
+    setRecentAction('move');
+    kernel.trigger('cut', [selected]);
+  };
+
+  const handlePaste = (e: React.SyntheticEvent) => {
+    kernel.trigger('paste', {
+      folders: clipboard,
+      directory: destination,
+    });
+  };
+
+  const handleMove = (e: React.SyntheticEvent) => {
+    kernel.trigger('move', {
+      folders: clipboard,
+      directory: destination,
+    });
+  };
+
+  const handleRename = (e: React.SyntheticEvent) => {
+    if (!selected) {
+      kernel.trigger('error', 'Select a Folder to rename');
+      return;
+    }
+
+    kernel.trigger('rename', {
+      name: '',
+      folder: selected,
+      directory: kernel.current,
+    });
+  };
+
+  const handleDelete = (e: React.SyntheticEvent) => {
+    if (!selected) {
+      kernel.trigger('error', 'Select a Folder to delete');
+      return;
+    }
+
+    kernel.trigger('delete', [selected]);
+  };
+
+  const handleClear = (e: React.SyntheticEvent) => {
+    kernel.trigger('cut', []);
+    setRecentAction(null);
+  };
+
+  const handleCopy = (e: React.SyntheticEvent) => {
+    setRecentAction('copy');
+    kernel.trigger('copy', [selected]);
+  };
+
+  const folderActions = {
+    handleCut,
+    handleCopy,
+    handleClear,
+    handleDelete,
+    handleMove,
+    handlePaste,
+    handleRename,
+    
+  };
+
+  const state = {
+    expanded,
+    recentAction,
+    setRecentAction,
+    setDestination,
+    handleActionExpand,
+    destination,
+  };
+
   return (
     <DashboardApiContext.Provider
       value={{
         clipboard,
+        folderActions,
+        state
       }} // Provide filesystem API functions as context
       {...rest}
     >
