+++ components/dashboard/index.tsx
@@ -1,3 +1,2 @@
 export * from './Dashboard';
-export * from './Context';
-export * from './Provider';
+export * from './context';
